rename dawsonv2 to localsidereal
